requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# LLM and RAG Orchestration
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
llama-index>=0.10.0

# Vector Storage
pinecone-client>=3.0.0
chromadb>=0.4.0

# Frontend
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Web Scraping and Data Processing
beautifulsoup4>=4.12.0
scrapy>=2.11.0
requests>=2.31.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# JSON-LD and Schema
rdflib>=7.0.0
jsonld>=1.0.0

# Utilities
python-dateutil>=2.8.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
